tests: add simple tests for utility commands
We depend on these commands working to help customers debug and for
various workflow, so let's add tests.

Test plan: make test
